chore: refactor homepage (move stats and article count to client component)
